Beginnings of a JSON API and front-end widget
Implements a basic JSON API for fetching threads and other things like
that.

Also includes a new plugin for displaying a widget on the front end of
your site listing out your tickets etc. Since this widget is entirely
Javascript powered, it doesn't even have to be on a WordPress blog to
work. Use it anywhere! :)