Don't attempt to send notifications when the `$to` address is empty.
We don't send notifications to the creater of a reply unless `supportflow_emails_notify_creator` is set to `true`, which can create a condition where the array of notification recipients is empty. In that case, we should abort instead of attempting to send a message to nobody, which will fail.