Add a new `supportflow_imap_folders` filter.
A lot of IMAP accounts are configured to namespace any mailboxes
with `INBOX.` and attempting to move to `ARCHIVE` or creating it
fails silently, causing the e-mails to be fetched over and over
again. With this filter, one is able to change the name of the
archive box.

This is a temporary solution until more user-friendly archive
mailbox picker is in place, and/or #245 is implemented.