Add required `get_columns()` method to list tables to prevent dying.
`Supportflow_Table` was introduced in 0064f68aba557302691583f2f0bc1ccdb7592a83, but was missing several required methods. There weren't any visible consequences to that until [wordpress 32644], which is slated for the upcoming 4.3 release.

`get_columns()` is now called whenever a table is displayed, and `WP_List_Table::get_columns()` will `die()` if it's not overwritten.