Add correct SMTP address to the From header.
Without this, the default address is used, which is an invalid address like wordpress@example.com. That can lead result in replies to messages sent via SupportFlow going to an invalid address.

That doesn't usually happen in real-world scenarios, because the sending MTA will correct the `From` header, but we should be setting it correctly anyway.